+++ packages/core/CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.2.1] - 2025-12-03
+
+### Changed
+- **Breaking:** Renamed `EMPTY_BLOC` to `EMPTY` in exports.
+
 ## [1.2.0] - 2025-12-03
 
 ### Changed